move firefly export code from visualization/firefly.py to data_objects/data_containers.py as a method. update docstring, guard import
